Refactor all advocate claim entities
Inline with addition of the supplementary claim type
alot of the logic for the claim entities can be shared.
